Bug reported by mar0x in forum post http://forum.unity3d.com/threads/uma-unity-multipurpose-avatar-on-the-asset-store.219175/page-50#post-2098457

+++ UMAProject/Assets/Standard Assets/UMA/Core/Scripts/SkinnedMeshCombiner.cs
@@ -267,7 +267,7 @@ namespace UMA
 			int dest = len1-1;
 			while (pos1 >= 0 && pos2 >= 0)
 			{
-				long i = mergedTransforms[pos1].hash - umaTransforms[pos2].hash;
+				long i = ((long)mergedTransforms[pos1].hash) - ((long)umaTransforms[pos2].hash);
 				if (i == 0)
 				{
 					mergedTransforms[dest] = mergedTransforms[pos1];
